remove the diann2mztab keep only the msstats.
